requirements.txt
google-cloud-firestore
fastapi
uvicorn
sqlalchemy